+++ cmd/demo/db_test.go
@@ -67,12 +67,7 @@ func TestLedgerSurvivesConnectionPool(t *testing.T) {
 	ds.mu.Lock()
 	defer ds.mu.Unlock()
 	for range 200 {
-		cust, pii := generateCustomer(ds.rng, ds.nextCustSeq, ds.products, ds.currentDay)
-		ds.nextCustSeq++
-		ds.persistCustomer(&cust, pii)
-		ds.customers = append(ds.customers, cust)
-		ds.addCustomerToLedger(&ds.customers[len(ds.customers)-1])
-		ds.fundCustomer(len(ds.customers) - 1)
+		ds.createCustomerLocked()
 	}
 
 	if len(ds.customers) != 200 {